requirements.txt
numpy==1.24.4
rasterio
torch==2.2.1
torchvision==0.17.1
transformers==4.39.3
Pillow
segment-anything
matplotlib